# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GPSLink Gateway",
        "description": "Available API for GPSLink Gateway",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/customer/ip-port/create": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Create Customer IP and Port",
                "operationId": "CreateCustomerIpPort",
                "requestBody": {
                    "description": "Customer IP and Port Information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "ip": {
                                        "type": "string"
                                    },
                                    "port": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "customer_id": 1,
                                    "ip": "123.123.123.123",
                                    "port": 123
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer IP and Port is successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerIpPorts"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "409": {
                        "description": "Customer IP and Port already exist!"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/customer/ip-port/ipPortById/{id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get Customer IP and Port by id",
                "operationId": "GetCustomerIpPortById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerIpPorts"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer IP and Port not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/customer/ip-port/list": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get list of Customer`s IP and Ports",
                "operationId": "CustomerIpPortList",
                "requestBody": {
                    "description": "Customer Id - NOTE: If customer_id object is omitted then all customers ip and port will be return.",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "transporter_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "customer_id": 0,
                                    "transporter_id": 0
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/customer/ip-port/update/{id}": {
            "put": {
                "tags": [
                    "Customer"
                ],
                "summary": "Updated Customer IP and Ports",
                "description": "Update Customer IP and Port information.",
                "operationId": "UpdateCustomerIPPORT",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated Customer IP and Port object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "ip": {
                                        "type": "string"
                                    },
                                    "port": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 0,
                                    "customer_id": 1,
                                    "ip": "123.123.123.123",
                                    "port": 123
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer IP and Port is successfully updated!"
                    },
                    "404": {
                        "description": "Customer IP and Port not found"
                    },
                    "409": {
                        "description": "Customer Id does not matched!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/customer/ip-port/delete/{id}": {
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "Delete Customer IP and Port by id",
                "operationId": "DeleteCustomerIpPort",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer IP and Port is successfully deleted!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerIpPorts"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer IP and Port does not exist!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomerIpPorts": {
                "title": "Customer IPs and Ports",
                "description": "Customer`s assigned IPs and Ports",
                "required": [
                    "customer_id",
                    "ip",
                    "port"
                ],
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "customer_id": {
                        "description": "Customer ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "ip": {
                        "description": "IP address",
                        "type": "string"
                    },
                    "port": {
                        "description": "Port",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}